+++ rummage/lib/gui/gui.py
@@ -15,6 +15,8 @@ from .controls.time_picker import TimePickerCtrl
 from .controls.result_lists import ResultFileList
 from .controls.result_lists import ResultContentList
 from .controls.encoding_list import EncodingList
+from .controls.file_picker import FilePickerAudioCtrl
+from .controls.file_picker import FilePickerCustomCtrl
 from .controls.load_search_list import SavedSearchList
 from .controls.search_chain_list import SearchChainList
 from .controls.search_error_list import ErrorList
@@ -935,6 +937,7 @@ class SettingsDialog ( wx.Dialog ):
         fgSizer52.Add( self.m_staticline91, 0, wx.EXPAND |wx.ALL, 5 )
 
         fgSizer54 = wx.FlexGridSizer( 0, 3, 0, 0 )
+        fgSizer54.AddGrowableCol( 2 )
         fgSizer54.SetFlexibleDirection( wx.BOTH )
         fgSizer54.SetNonFlexibleGrowMode( wx.FLEX_GROWMODE_SPECIFIED )
 
@@ -945,7 +948,7 @@ class SettingsDialog ( wx.Dialog ):
         fgSizer54.Add( self.m_prerelease_checkbox, 0, wx.ALL, 5 )
 
         self.m_check_update_button = wx.Button( self.m_general_panel, wx.ID_ANY, u"Check Now", wx.DefaultPosition, wx.DefaultSize, 0 )
-        fgSizer54.Add( self.m_check_update_button, 0, wx.ALL, 5 )
+        fgSizer54.Add( self.m_check_update_button, 0, wx.ALL|wx.ALIGN_RIGHT, 5 )
 
 
         fgSizer52.Add( fgSizer54, 1, wx.EXPAND, 5 )
@@ -1128,19 +1131,30 @@ class SettingsDialog ( wx.Dialog ):
 
         self.m_visual_alert_checkbox = wx.CheckBox( self.m_notify_panel, wx.ID_ANY, u"Notification popup", wx.DefaultPosition, wx.DefaultSize, 0 )
         self.m_visual_alert_checkbox.SetValue(True)
-        fgSizer35.Add( self.m_visual_alert_checkbox, 0, wx.ALL, 5 )
+        fgSizer35.Add( self.m_visual_alert_checkbox, 0, wx.ALL|wx.ALIGN_CENTER_VERTICAL, 5 )
 
         m_notify_choiceChoices = []
         self.m_notify_choice = wx.Choice( self.m_notify_panel, wx.ID_ANY, wx.DefaultPosition, wx.DefaultSize, m_notify_choiceChoices, 0 )
         self.m_notify_choice.SetSelection( 0 )
         fgSizer35.Add( self.m_notify_choice, 0, wx.ALIGN_RIGHT|wx.ALL, 5 )
 
-        self.m_audio_alert_checkbox = wx.CheckBox( self.m_notify_panel, wx.ID_ANY, u"Alert Sound", wx.DefaultPosition, wx.DefaultSize, 0 )
+        self.m_audio_alert_checkbox = wx.CheckBox( self.m_notify_panel, wx.ID_ANY, u"Alert sound", wx.DefaultPosition, wx.DefaultSize, 0 )
         self.m_audio_alert_checkbox.SetValue(True)
-        fgSizer35.Add( self.m_audio_alert_checkbox, 0, wx.ALL, 5 )
+        fgSizer35.Add( self.m_audio_alert_checkbox, 0, wx.ALL|wx.ALIGN_CENTER_VERTICAL, 5 )
 
+        self.m_sound_picker = FilePickerAudioCtrl(self.m_notify_panel, wx.ID_ANY, wx.EmptyString, u"Select audio file", u"*.*", wx.DefaultPosition, wx.DefaultSize, wx.FLP_DEFAULT_STYLE|wx.FLP_FILE_MUST_EXIST)
+        self.m_sound_picker.Bind( wx.EVT_FILEPICKER_CHANGED, self.on_sound_change)
+        fgSizer35.Add( self.m_sound_picker, 0, wx.ALL|wx.EXPAND|wx.ALIGN_CENTER_VERTICAL, 5 )
 
-        fgSizer35.Add( ( 0, 0), 1, wx.EXPAND, 5 )
+        self.m_sound_player_label = wx.StaticText( self.m_notify_panel, wx.ID_ANY, u"Alert player", wx.DefaultPosition, wx.DefaultSize, 0 )
+        self.m_sound_player_label.Wrap( -1 )
+
+        fgSizer35.Add( self.m_sound_player_label, 0, wx.ALL, 5 )
+
+        m_sound_player_choiceChoices = []
+        self.m_sound_player_choice = wx.Choice( self.m_notify_panel, wx.ID_ANY, wx.DefaultPosition, wx.DefaultSize, m_sound_player_choiceChoices, 0 )
+        self.m_sound_player_choice.SetSelection( 0 )
+        fgSizer35.Add( self.m_sound_player_choice, 0, wx.ALL|wx.ALIGN_RIGHT, 5 )
 
         self.m_term_note_label = wx.StaticText( self.m_notify_panel, wx.ID_ANY, u"Path to terminal-notifier", wx.DefaultPosition, wx.DefaultSize, 0 )
         self.m_term_note_label.Wrap( -1 )
@@ -1149,10 +1163,9 @@ class SettingsDialog ( wx.Dialog ):
 
         fgSizer35.Add( self.m_term_note_label, 0, wx.ALL|wx.ALIGN_CENTER_VERTICAL, 5 )
 
-        self.m_term_note_picker = wx.FilePickerCtrl( self.m_notify_panel, wx.ID_ANY, wx.EmptyString, u"Select a file", u"*.*", wx.DefaultPosition, wx.DefaultSize, wx.FLP_DEFAULT_STYLE|wx.FLP_FILE_MUST_EXIST )
-        self.m_term_note_picker.Enable( False )
-
-        fgSizer35.Add( self.m_term_note_picker, 0, wx.ALL|wx.ALIGN_CENTER_VERTICAL|wx.EXPAND, 5 )
+        self.m_term_note_picker = FilePickerCustomCtrl(self.m_notify_panel, wx.ID_ANY, wx.EmptyString, u"Select file", u"*", wx.DefaultPosition, wx.DefaultSize, wx.FLP_DEFAULT_STYLE|wx.FLP_FILE_MUST_EXIST)
+        self.m_term_note_picker.Bind( wx.EVT_FILEPICKER_CHANGED, self.on_term_note_change)
+        fgSizer35.Add( self.m_term_note_picker, 0, wx.ALL|wx.EXPAND|wx.ALIGN_CENTER_VERTICAL, 5 )
 
 
         fgSizer35.Add( ( 0, 0), 1, wx.EXPAND, 5 )
@@ -1303,7 +1316,7 @@ class SettingsDialog ( wx.Dialog ):
         self.m_visual_alert_checkbox.Bind( wx.EVT_CHECKBOX, self.on_notify_toggle )
         self.m_notify_choice.Bind( wx.EVT_CHOICE, self.on_notify_choice )
         self.m_audio_alert_checkbox.Bind( wx.EVT_CHECKBOX, self.on_alert_toggle )
-        self.m_term_note_picker.Bind( wx.EVT_FILEPICKER_CHANGED, self.on_term_note_change )
+        self.m_sound_player_choice.Bind( wx.EVT_CHOICE, self.on_player_selected )
         self.m_notify_test_button.Bind( wx.EVT_BUTTON, self.on_notify_test_click )
         self.m_history_clear_button.Bind( wx.EVT_BUTTON, self.on_clear_history )
         self.m_back_ext_textbox.Bind( wx.EVT_TEXT, self.on_back_ext_changed )
@@ -1393,7 +1406,7 @@ class SettingsDialog ( wx.Dialog ):
     def on_alert_toggle( self, event ):
         event.Skip()
 
-    def on_term_note_change( self, event ):
+    def on_player_selected( self, event ):
         event.Skip()
 
     def on_notify_test_click( self, event ):