+++ docs/src/markdown/changelog.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## Current
+## 4.7.0 (Development)
 
 - **NEW**: Add `col0` variable for editor configuration to allow for using a zero based column value instead of one based column value for editors that require it.
 - **NEW**: Add global option to turn off alternating row colors in lists.
-- **FIX**: Better attempt to get appropriate creation on all Linux systems.
+- **NEW**: Provide support for different players on Linux: `paplay`, `aplay`, and `play` (`sox`). Supported sound formats vary.
+- **NEW**: Provide interface for selecting custom notification sounds.
+- **FIX**: Better attempt to get appropriate file creation time on all Linux systems.
 - **FIX**: Better fix for time picker getting wrong background.
 - **FIX**: Ensure column sizing includes header size.
 - **FIX**: When showing timed status event for results in status bar (full path name), decrease length of time that the temporary status is shown and clear temporary status when the cursor leaves result items.